pyproject.toml
[project]
name = "dinesmartai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.28.0",
    "google-maps-places>=0.8.0",
    "googlemaps>=4.10.0",
    "httpx>=0.28.1",
]
